+++ koopovereenkomst-v3-simple/src/ui-components/MinimalizationModal.tsx
